+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Bump GH Actions `upload-artifact` to v3
 - Bump logback to v1.2.11
 - Move wildcard formatting to search query
+- Update GH Actions `checkout` to v3
 - [In Progress: 17 Feb 2022] Add `CriticalAppUpdateDialog`
 
 ### Changes